+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+v1.5.5
+- Updated plugins
+
+v1.5.4
+- Added jshint auditing back
+- Updated npm packages
+
+v1.5.3
+- Added jandlebars, html-format, jshint-webpack-plugin npm packages
+
 v1.5.2
 - Added a11y-webpack-plugin back in